GRT-63 [MOVE] 폴더 컨벤션 구조에 맞게 컴포넌트,스토리 이동
